requirements.txt
numpy==1.16.2
opencv-python==4.0.0.21
pytesseract==0.2.6
flask==1.0.2
werkzeug==0.14.1
pillow==5.4.1